+++ pytorch_lightning/plugins/__init__.py
@@ -20,10 +20,10 @@ from pytorch_lightning.plugins.precision.sharded_native_amp import ShardedNative
 from pytorch_lightning.plugins.precision.tpu import TPUPrecisionPlugin
 from pytorch_lightning.plugins.precision.tpu_bf16 import TPUBf16PrecisionPlugin
 from pytorch_lightning.plugins.training_type.ddp import DDPStrategy
-from pytorch_lightning.plugins.training_type.ddp2 import DDP2Plugin
+from pytorch_lightning.plugins.training_type.ddp2 import DDP2Strategy
 from pytorch_lightning.plugins.training_type.ddp_spawn import DDPSpawnPlugin
 from pytorch_lightning.plugins.training_type.deepspeed import DeepSpeedStrategy
-from pytorch_lightning.plugins.training_type.dp import DataParallelPlugin
+from pytorch_lightning.plugins.training_type.dp import DataParallelStrategy
 from pytorch_lightning.plugins.training_type.fully_sharded import DDPFullyShardedStrategy
 from pytorch_lightning.plugins.training_type.horovod import HorovodStrategy
 from pytorch_lightning.plugins.training_type.ipu import IPUStrategy
@@ -31,7 +31,7 @@ from pytorch_lightning.plugins.training_type.parallel import ParallelPlugin
 from pytorch_lightning.plugins.training_type.sharded import DDPShardedStrategy
 from pytorch_lightning.plugins.training_type.sharded_spawn import DDPSpawnShardedPlugin
 from pytorch_lightning.plugins.training_type.single_device import SingleDeviceStrategy
-from pytorch_lightning.plugins.training_type.single_tpu import SingleTPUPlugin
+from pytorch_lightning.plugins.training_type.single_tpu import SingleTPUStrategy
 from pytorch_lightning.plugins.training_type.tpu_spawn import TPUSpawnStrategy
 from pytorch_lightning.plugins.training_type.training_type_plugin import Strategy
 
@@ -43,8 +43,8 @@ __all__ = [
     "TorchCheckpointIO",
     "XLACheckpointIO",
     "ApexMixedPrecisionPlugin",
-    "DataParallelPlugin",
-    "DDP2Plugin",
+    "DataParallelStrategy",
+    "DDP2Strategy",
     "DDPStrategy",
     "DDPSpawnPlugin",
     "DDPFullyShardedStrategy",
@@ -59,7 +59,7 @@ __all__ = [
     "ShardedNativeMixedPrecisionPlugin",
     "FullyShardedNativeMixedPrecisionPlugin",
     "SingleDeviceStrategy",
-    "SingleTPUPlugin",
+    "SingleTPUStrategy",
     "TPUPrecisionPlugin",
     "TPUBf16PrecisionPlugin",
     "TPUSpawnStrategy",