+++ tests/accelerators/test_accelerator_connector.py
@@ -25,8 +25,8 @@ from pytorch_lightning.accelerators.accelerator import Accelerator
 from pytorch_lightning.accelerators.cpu import CPUAccelerator
 from pytorch_lightning.accelerators.gpu import GPUAccelerator
 from pytorch_lightning.plugins import (
-    DataParallelPlugin,
-    DDP2Plugin,
+    DataParallelStrategy,
+    DDP2Strategy,
     DDPShardedStrategy,
     DDPSpawnPlugin,
     DDPSpawnShardedPlugin,
@@ -128,7 +128,7 @@ def test_accelerator_choice_ddp2_slurm(*_):
         trainer = Trainer(fast_dev_run=True, accelerator="ddp2", gpus=2)
     assert trainer._accelerator_connector._is_slurm_managing_tasks()
     assert isinstance(trainer.accelerator, GPUAccelerator)
-    assert isinstance(trainer.training_type_plugin, DDP2Plugin)
+    assert isinstance(trainer.training_type_plugin, DDP2Strategy)
     assert isinstance(trainer.training_type_plugin.cluster_environment, SLURMEnvironment)
     assert trainer.training_type_plugin.cluster_environment.local_rank() == 1
     assert trainer.training_type_plugin.local_rank == 1
@@ -176,7 +176,7 @@ def test_accelerator_choice_ddp2_te(*_):
     with pytest.deprecated_call(match=r"accelerator='ddp2'\)` has been deprecated in v1.5"):
         trainer = Trainer(fast_dev_run=True, accelerator="ddp2", gpus=2)
     assert isinstance(trainer.accelerator, GPUAccelerator)
-    assert isinstance(trainer.training_type_plugin, DDP2Plugin)
+    assert isinstance(trainer.training_type_plugin, DDP2Strategy)
     assert isinstance(trainer.training_type_plugin.cluster_environment, TorchElasticEnvironment)
     assert trainer.training_type_plugin.cluster_environment.local_rank() == 1
     assert trainer.training_type_plugin.local_rank == 1
@@ -627,8 +627,8 @@ def test_strategy_choice_cpu_plugin(tmpdir, plugin):
         ("ddp_spawn_find_unused_parameters_false", DDPSpawnPlugin),
         ("ddp", DDPStrategy),
         ("ddp_find_unused_parameters_false", DDPStrategy),
-        ("ddp2", DDP2Plugin),
-        ("dp", DataParallelPlugin),
+        ("ddp2", DDP2Strategy),
+        ("dp", DataParallelStrategy),
         ("ddp_sharded", DDPShardedStrategy),
         ("ddp_sharded_spawn", DDPSpawnShardedPlugin),
         pytest.param("deepspeed", DeepSpeedStrategy, marks=RunIf(deepspeed=True)),
@@ -733,12 +733,12 @@ def test_strategy_choice_ddp_slurm(setup_distributed_mock, strategy):
 @mock.patch("torch.cuda.set_device")
 @mock.patch("torch.cuda.device_count", return_value=2)
 @mock.patch("pytorch_lightning.plugins.DDPStrategy.setup_distributed", autospec=True)
-@pytest.mark.parametrize("strategy", ["ddp2", DDP2Plugin()])
+@pytest.mark.parametrize("strategy", ["ddp2", DDP2Strategy()])
 def test_strategy_choice_ddp2_slurm(set_device_mock, device_count_mock, setup_distributed_mock, strategy):
     trainer = Trainer(fast_dev_run=True, strategy=strategy, gpus=2)
     assert trainer._accelerator_connector._is_slurm_managing_tasks()
     assert isinstance(trainer.accelerator, GPUAccelerator)
-    assert isinstance(trainer.training_type_plugin, DDP2Plugin)
+    assert isinstance(trainer.training_type_plugin, DDP2Strategy)
     assert isinstance(trainer.training_type_plugin.cluster_environment, SLURMEnvironment)
     assert trainer.training_type_plugin.cluster_environment.local_rank() == 1
     assert trainer.training_type_plugin.local_rank == 1
@@ -784,7 +784,7 @@ def test_strategy_choice_ddp_te(*_):
 def test_strategy_choice_ddp2_te(*_):
     trainer = Trainer(fast_dev_run=True, strategy="ddp2", gpus=2)
     assert isinstance(trainer.accelerator, GPUAccelerator)
-    assert isinstance(trainer.training_type_plugin, DDP2Plugin)
+    assert isinstance(trainer.training_type_plugin, DDP2Strategy)
     assert isinstance(trainer.training_type_plugin.cluster_environment, TorchElasticEnvironment)
     assert trainer.training_type_plugin.cluster_environment.local_rank() == 1
     assert trainer.training_type_plugin.local_rank == 1
@@ -879,11 +879,11 @@ def test_unsupported_tpu_choice(monkeypatch):
     with pytest.raises(MisconfigurationException, match=r"accelerator='tpu', precision=64\)` is not implemented"):
         Trainer(accelerator="tpu", precision=64)
 
-    with pytest.raises(ValueError, match="TPUAccelerator` can only be used with a `SingleTPUPlugin`"):
+    with pytest.raises(ValueError, match="TPUAccelerator` can only be used with a `SingleTPUStrategy`"):
         with pytest.warns(UserWarning, match=r"accelerator='tpu', precision=16\)` but native AMP is not supported"):
             Trainer(accelerator="tpu", precision=16)
 
-    with pytest.raises(ValueError, match="TPUAccelerator` can only be used with a `SingleTPUPlugin`"):
+    with pytest.raises(ValueError, match="TPUAccelerator` can only be used with a `SingleTPUStrategy`"):
         with pytest.warns(UserWarning, match=r"accelerator='tpu', precision=16\)` but apex AMP is not supported"):
             Trainer(accelerator="tpu", precision=16, amp_backend="apex")